code_tf1_13/requirements.txt
tensorflow-gpu==1.12.0
coloredlogs
h5py
numpy
opencv-python
ujson
